feat: Add node-html-parser package for HTML parsing
This commit adds the `node-html-parser` package to the project's dependencies. This package is used for parsing HTML in the application. It is required for the newly added signup, signin, and dashboard pages, as well as for other components and services that handle HTML parsing.

Please review and merge this change.